backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
google-genai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
